+++ lib/tests/unit/transforms/test_decorators.py
@@ -188,6 +188,87 @@ def test_thread_limiting(
     mock_semaphore.return_value.release.assert_called_once_with()
 
 
+@pytest.mark.parametrize(
+    "max_thread_count,patch_str",
+    (
+        (None, "threading.BoundedSemaphore"),
+        (_thread_limiter.ThreadLimit.DEFAULT, "threading.BoundedSemaphore"),
+        (_thread_limiter.ThreadLimit.NONE, "_DummySemaphore"),
+    ),
+)
+def test_thread_limiting_process_method(
+    max_thread_count, patch_str, kmsg, mock_config, mocker, monkeypatch
+):
+    mock_function = mocker.Mock()
+    mock_semaphore = mocker.Mock()
+    monkeypatch.setattr(
+        f"klio.utils._thread_limiter.{patch_str}", mock_semaphore
+    )
+
+    kwargs = {}
+    if max_thread_count is not None:
+        kwargs["max_thread_count"] = max_thread_count
+
+    class TestDoFn(beam.DoFn):
+        @decorators._handle_klio(**kwargs)
+        def process(self, msg_data):
+            mock_semaphore.return_value.acquire.assert_called_once_with()
+            mock_semaphore.return_value.release.assert_not_called()
+            mock_function()
+            return msg_data
+
+    dofn = TestDoFn()
+
+    # notice - process is always a generator, even when `return` is used
+    gen = dofn.process(kmsg.SerializeToString())
+
+    # need to pull everything from the generator
+    [z for z in gen]
+
+    assert 1 == mock_function.call_count
+    mock_semaphore.return_value.acquire.assert_called_once_with()
+    mock_semaphore.return_value.release.assert_called_once_with()
+
+
+@pytest.mark.parametrize(
+    "max_thread_count,patch_str",
+    (
+        (None, "threading.BoundedSemaphore"),
+        (_thread_limiter.ThreadLimit.DEFAULT, "threading.BoundedSemaphore"),
+        (_thread_limiter.ThreadLimit.NONE, "_DummySemaphore"),
+    ),
+)
+def test_thread_limiting_nonprocess_method(
+    max_thread_count, patch_str, kmsg, mock_config, mocker, monkeypatch
+):
+    # Unlike the `process` method, non-process method should be treated as a
+    # regular function
+    mock_function = mocker.Mock()
+    mock_semaphore = mocker.Mock()
+    monkeypatch.setattr(
+        f"klio.utils._thread_limiter.{patch_str}", mock_semaphore
+    )
+
+    kwargs = {}
+    if max_thread_count is not None:
+        kwargs["max_thread_count"] = max_thread_count
+
+    class TestMethod(object):
+        @decorators._handle_klio(**kwargs)
+        def other_method(self, msg_data):
+            mock_semaphore.return_value.acquire.assert_called_once_with()
+            mock_semaphore.return_value.release.assert_not_called()
+            mock_function()
+            return msg_data
+
+    dofn = TestMethod()
+    dofn.other_method(kmsg.SerializeToString())
+
+    assert 1 == mock_function.call_count
+    mock_semaphore.return_value.acquire.assert_called_once_with()
+    mock_semaphore.return_value.release.assert_called_once_with()
+
+
 def test_thread_limiting_custom_limiter(
     kmsg, mock_config, mocker, monkeypatch
 ):
@@ -408,3 +489,54 @@ def test_serialize_klio_metrics(mock_config, kmsg):
     assert len(pcoll) == msg_timer.committed.count
     assert "simple_map" == msg_timer.key.metric.namespace
     assert "kmsg-timer" == msg_timer.key.metric.name
+
+
+@pytest.mark.parametrize(
+    "items",
+    ([], [1], [1, 2], [1, 2, 3], [Exception("Hey")], [1, Exception("HEY")]),
+)
+def test_threadlimitgenerator(mocker, items):
+    def test_gen(items):
+        for item in items:
+            if isinstance(item, Exception):
+                raise item
+            yield item
+
+    limiter = mocker.Mock()
+
+    gen = decorators.ThreadLimitGenerator(limiter, test_gen(items))
+
+    limiter.acquire.assert_called_once_with()
+    limiter.release.assert_not_called()
+    for item in items:
+        if isinstance(item, Exception):
+            with pytest.raises(Exception):
+                next(gen)
+            limiter.release.assert_called_once_with()
+            break
+        else:
+            next(gen)
+            limiter.release.assert_not_called()
+
+    with pytest.raises(StopIteration):
+        next(gen)
+
+    limiter.release.assert_called_once_with()
+
+
+def test_threadlimitgenerator_del_release(mocker):
+    # This test verifies the __del__ method properly releases the semaphore
+
+    limiter = mocker.Mock()
+
+    def subfunc():
+        _ = decorators.ThreadLimitGenerator(limiter, iter([1, 2]))
+
+        limiter.acquire.assert_called_once_with()
+        limiter.release.assert_not_called()
+
+    subfunc()
+    limiter.release.assert_called_once_with()
+
+
+patcher.stop()